Fix: Check for 'gameId' in localStorage in both 'index.js' and 'user-controller-events.js'
With this fix, if the user has no 'gameId' in localStorage,
the app will request for a new one and store it in localStorage.

If user deletes the 'gameId' from localStorage, and attempts to refresh the page,
or refresh the scores table, the app will request for a new one and store it
in localStorage.